pyproject.toml
[project]
name = "ddb-ops"
version = "0.1.0"
description = "Testing out various access patterns for DynamoDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "basedpyright>=1.27.0",
    "boto3>=1.36.13",
    "boto3-stubs[dynamodb]>=1.36.13",
    "pynamodb>=6.0.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.15.1",
]

# BELOW ARE REQUIRED FOR PACKAGES
[tool.hatch.build.targets.wheel]
packages = ["src/ddb_ops"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "moto[dynamodb,dynamodb2]>=5.1.0",
    "pytest>=8.3.4",
    "pytest-env>=1.1.5",
    "ruff>=0.9.6",
]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
include = ["boto3-stubs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
